+++ app/mailers/integral/contact_mailer.rb
@@ -50,6 +50,7 @@ module Integral
     # Adds the enquiry subject if it exists otherwise uses only reference
     def forwarding_subject(enquiry)
       return "#{enquiry.reference} #{I18n.t('integral.contact_mailer.forward_enquiry.subject')}" if enquiry.subject.blank?
+
       "#{enquiry.reference} #{I18n.t('integral.contact_mailer.forward_enquiry.subject')} - #{enquiry.subject}"
     end
   end