+++ lib/generators/templates/carrierwave.rb
@@ -25,9 +25,9 @@ CarrierWave.configure do |config|
   config.aws_bucket = ENV.fetch('AWS_S3_BUCKET_NAME')
   config.aws_acl    = 'public-read'
   config.aws_credentials = {
-    access_key_id:     ENV.fetch('AWS_ACCESS_KEY_ID'),
+    access_key_id: ENV.fetch('AWS_ACCESS_KEY_ID'),
     secret_access_key: ENV.fetch('AWS_SECRET_ACCESS_KEY'),
-    region:            ENV.fetch('AWS_REGION') # Required
+    region: ENV.fetch('AWS_REGION') # Required
   }
 
   # CDN configuration (Recommended)