+++ lib/integral/engine.rb
@@ -73,7 +73,7 @@ module Integral
     # Allows engine factories to be reused by application
     initializer 'model_core.factories', after: 'factory_bot.set_factory_paths' do
       if defined?(FactoryBot)
-        FactoryBot.definition_file_paths << File.expand_path('../../../spec/factories', __FILE__)
+        FactoryBot.definition_file_paths << File.expand_path('../../spec/factories', __dir__)
       end
     end