+++ spec/dummy/config/application.rb
@@ -9,6 +9,6 @@ require "integral"
 
 module Dummy
   class Application < Rails::Application
-    config.active_record.sqlite3.represent_boolean_as_integer = true
+    config.action_dispatch.return_only_media_type_on_content_type = false
   end
 end