+++ spec/rails_helper.rb
@@ -23,6 +23,7 @@ require 'capybara/poltergeist'
 require 'paper_trail/frameworks/rspec'
 
 Capybara.default_driver = :poltergeist
+Capybara.server = :webrick
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are