+++ CHANGELOG.md
@@ -7,18 +7,18 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Breaking Changes
 
--
+- Major Rails upgrade & upgraded major versions for acts-as-taggable-on, Paper Trail, rspec-rails, Devise, CoffeeScript, Client Side Validations, Haml-Rails, Factory Bot, Capybara & Rubocop
 
 ### Added
 
--
+- Rails 6!
 
 ### Fixed
 
 -
 
 ### Dependancies
--
+- Upgraded major versions for acts-as-taggable-on, Paper Trail, rspec-rails, Devise, CoffeeScript, Client Side Validations, Haml-Rails, Factory Bot, Capybara & Rubocop
 
 ## v1.3.0 - 2019-09-14